requirements.txt
streamlit
openai
langchain
langsmith
python-dotenv
aiohttp
